// src/data/data.dto.ts
export class DataCreateDto {
  hash: string; // unique hash for the data (identifier)
  data: string; // encrypted data
  type: string; // generic type for mobile handling, eg: 'kyc'
}

export class DataResponseDto {
  hash: string; // unique hash for the data (identifier)
  secureNonce: string; // secureNonce for the data (that was used to create the secureHash)
}

export class DataFetchDto {
  hash: string; // unique hash for the data (identifier)
}

export class DataUpdateDto {
  hash: string; // unique hash for the data (identifier)
  secureHash: string; // secure hash generated by sdk with secureNonce
  success: boolean; // success status of the data handling (passing)
  response: string; // any response message from the data handling (or error message)
}